requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0
# Optional dependencies for Azure Table Storage
azure-data-tables>=12.0.0
azure-identity>=1.10.0

# Test dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
httpx>=0.23.0